Exit 1
I don't think `! COMMAND` is marked as failure (in the sense of it will
cause the script to exit when run with `set -e`) if COMMAND was
successful.

+++ tests/missing-file/run.sh
@@ -1,3 +1,3 @@
-! "$MILLET" does-not-exist.sml >out.tmp
+if "$MILLET" does-not-exist.sml >out.tmp; then exit 1; fi
 diff out.txt out.tmp
 rm out.tmp
